% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{allrow_cor}
\alias{allrow_cor}
\title{computing correlations between all rows in two matrices}
\usage{
allrow_cor(A, B)
}
\arguments{
\item{A}{first matrix}

\item{B}{second matrix - should be exactly same dimension as A}
}
\description{
Efficient version for computing correlations between all rows
}
