% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset.r
\name{gset_add_genes}
\alias{gset_add_genes}
\title{Add specific genes to the set}
\usage{
gset_add_genes(gset, genes, subset_id)
}
\arguments{
\item{gset}{gene set object}

\item{genes}{genes to add}

\item{subset_id}{subset id to which the new genes will be assigned}
}
\description{
Add specific genes to the set
}
