% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset.r
\name{gset_get_feat_mat}
\alias{gset_get_feat_mat}
\title{extract umi matrix for the genes in the set, possibly donwsampling}
\usage{
gset_get_feat_mat(gset_id, mat_id, downsamp = F)
}
\arguments{
\item{gset_id}{gene set in scdb}

\item{mat_id}{mat in scdb}

\item{downsamp}{if this is true the returned matrix is downsampled}
}
\description{
extract umi matrix for the genes in the set, possibly donwsampling
}
