% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset.r
\name{gset_import_table}
\alias{gset_import_table}
\title{Import gene set for a text table}
\usage{
gset_import_table(fn, desc = NULL)
}
\arguments{
\item{fn}{file name to read from}

\item{desc}{tetual description of the gset, if null the desc will be the file name}
}
\description{
Import gene set for a text table
}
