% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset.r
\name{gset_new_gset}
\alias{gset_new_gset}
\title{Generating a new gset}
\usage{
gset_new_gset(sets, desc)
}
\arguments{
\item{sets}{a vector where names are genes and values are set ids}

\item{desc}{tetual description of the gset}
}
\description{
Generating a new gset
}
