% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset.r
\name{gset_new_restrict_gset}
\alias{gset_new_restrict_gset}
\title{Generate a new gene set from an existing one, filtered by a list of genes}
\usage{
gset_new_restrict_gset(gset, filt_gset, inverse = F, desc)
}
\arguments{
\item{gset}{a gene set object}

\item{filt_gset}{gset of genes to filter by}

\item{inverse}{if true, omit the genes in the genes parameter}

\item{desc}{description of the new gene set}
}
\description{
Generate a new gene set from an existing one, filtered by a list of genes
}
