% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset.r
\name{gset_write_table}
\alias{gset_write_table}
\title{Exprt gene set to a table}
\usage{
gset_write_table(gset, fn)
}
\arguments{
\item{gset}{gene set}

\item{fn}{file name to save to}
}
\description{
Exprt gene set to a table
}
