% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.r
\name{initialize,tgCellGraph-method}
\alias{initialize,tgCellGraph-method}
\title{This constructs graph object from a data frame on columns col1,col2, weight, (the format returened by tgs_cor_graph). Note this is not intended to serve natively complex graph algs etc, but is just a container for use by scdb and operaitons on scmats (creating a graph) and mcell (using the graph to compute graph covers)}
\usage{
\S4method{initialize}{tgCellGraph}(.Object, gr, cnames)
}
\arguments{
\item{gr}{data frame with columns col1 (first node), col2 (second node), weight - float between 0 and 1.}
}
\description{
This constructs graph object from a data frame on columns col1,col2, weight, (the format returened by tgs_cor_graph). Note this is not intended to serve natively complex graph algs etc, but is just a container for use by scdb and operaitons on scmats (creating a graph) and mcell (using the graph to compute graph covers)
}
