% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclust.r
\name{initialize,tgCoClust-method}
\alias{initialize,tgCoClust-method}
\title{Construct a coclust object}
\usage{
\S4method{initialize}{tgCoClust}(.Object, graph_id, coclust, n_samp)
}
\arguments{
\item{graph_id}{just cahce the grpah id}

\item{coclust}{data frame with fields node1, node2, count}

\item{n_samp}{vector with statistics on number of time each node was sampled}
}
\description{
Construct a coclust object
}
