% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset.r
\name{initialize,tgGeneSets-method}
\alias{initialize,tgGeneSets-method}
\title{tgGeneSets public constructor}
\usage{
\S4method{initialize}{tgGeneSets}(.Object, sets, desc = NULL)
}
\arguments{
\item{sets}{a named vector assigning set id to genes}

\item{desc}{an optional string describing the gene set}
}
\description{
This constructs a tgGene sets object, nothing fancy
}
