% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc2d.r
\docType{methods}
\name{initialize,tgMC2D-method}
\alias{initialize,tgMC2D-method}
\title{Construct a meta cell 2D embedding}
\usage{
\S4method{initialize}{tgMC2D}(.Object, mc_id, mc_x, mc_y, sc_x, sc_y, graph)
}
\arguments{
\item{mc_x}{x coordinates of metacells}

\item{mc_y}{y coordinates of metacells}

\item{sc_x}{x coordinates of cells}

\item{sc_y}{y coordinates of cells}

\item{graph}{data.frame with fields mc1, mc2 definining the graph}

\item{mc}{metacell object}
}
\description{
Just filling up the slots, actuall computation is done by specific algorithms (mcell_mc2d_force)
}
