% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.r
\name{initialize,tgMCAtlas-method}
\alias{initialize,tgMCAtlas-method}
\title{Construct a meta cell reference atlas}
\usage{
\S4method{initialize}{tgMCAtlas}(.Object, mat_id, naming_type, mc_id, graph_id, gset_id, type_annot)
}
\arguments{
\item{mat_id}{umi matrix of the reference}

\item{naming_type}{gene naming type}

\item{mc_id}{metacell object id}

\item{graph_id}{KNN adjacencies for the atlas}

\item{gset_id}{gene set id for defining features}

\item{type_annot}{data frame with type_id, type_name, type_color fields}
}
\description{
Construct a meta cell reference atlas
}
