% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.r
\docType{methods}
\name{initialize,tgMCCov-method}
\alias{initialize,tgMCCov-method}
\title{Construct a meta cell object}
\usage{
\S4method{initialize}{tgMCCov}(.Object, mc, outliers = c(), scmat)
}
\arguments{
\item{mc}{assignment of metacell id to cell}

\item{scmat}{a single cell RNA matrix object}
}
\description{
This constructs a meta cell cover object. It gets an MC assignment (cell->MC_ID), and a matrix, and call standard api of this class to compute the footprints.
}
