% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.r
\name{initialize,tgMCManifGraph-method}
\alias{initialize,tgMCManifGraph-method}
\title{Construct a meta cell manifold graph}
\usage{
\S4method{initialize}{tgMCManifGraph}(.Object, mc_id, mgraph)
}
\arguments{
\item{mc_id}{metacell object id}

\item{mgraph}{data fra,e defining mc1, mc2, distance}
}
\description{
Construct a meta cell manifold graph
}
