% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\name{initialize,tgMCTNetwork-method}
\alias{initialize,tgMCTNetwork-method}
\title{Construct a meta cell time network}
\usage{
\S4method{initialize}{tgMCTNetwork}(.Object, mc_id, mgraph_id, cell_time)
}
\arguments{
\item{mc_id}{metacell object id}

\item{mgraph_id}{metacell graph object id}

\item{cell_time}{vector assigning a time (number) to each cell}

\item{mc_prolif}{factor per mc}
}
\description{
Construct a meta cell time network
}
