% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_colorize.r
\name{mc_colorize_default}
\alias{mc_colorize_default}
\title{colorize metacell using an ugly default color spectrum, or a user supplied one}
\usage{
mc_colorize_default(mc_id, spectrum = NULL)
}
\arguments{
\item{mc_id}{metacell id in scdb}
}
\description{
colorize metacell using an ugly default color spectrum, or a user supplied one
}
