% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_colorize.r
\name{mc_colorize_from_ref_mc}
\alias{mc_colorize_from_ref_mc}
\title{colorize metacell by projecting colors from another metacell on a similar (not identical) set of cells}
\usage{
mc_colorize_from_ref_mc(mc_id, mc_src_id, min_color_frac = 0.5)
}
\arguments{
\item{mc_id}{metacell id in scdb}

\item{mc_src_id}{the metacell to use as a refernece}

\item{min_color_frac}{minimal fraction of cells with a given color in order to perform color projection.}
}
\description{
colorize metacell by projecting colors from another metacell on a similar (not identical) set of cells
}
