% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_colorize.r
\name{mc_colorize_sup_hierarchy}
\alias{mc_colorize_sup_hierarchy}
\title{colorize metacells using a set of super MCs derived by hclust, colored according to a user defined table}
\usage{
mc_colorize_sup_hierarchy(mc_id, supmc, supmc_key, gene_key = NULL)
}
\arguments{
\item{mc_id}{metacell id in scdb}

\item{supmc}{output from mcell_mc_hierarchy, defining groups of mcs}

\item{supmc_key}{filename of color index, specifying an ordered list of supmc ids and colors, to be colored in the order of apperance (i.e. last line overrid previous lines). Expected field names: supid, color, name}
}
\description{
colorize metacells using a set of super MCs derived by hclust, colored according to a user defined table
}
