% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.r
\name{mc_compute_cov_gc}
\alias{mc_compute_cov_gc}
\title{Compute fraction of non zero expressing cells per gene and mc}
\usage{
mc_compute_cov_gc(mc, us)
}
\arguments{
\item{mc}{a metacell object}

\item{us}{umi matrix}
}
\description{
Compute fraction of non zero expressing cells per gene and mc
}
