% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.r
\name{mc_compute_fp}
\alias{mc_compute_fp}
\title{Compute metacell gene footprint}
\usage{
mc_compute_fp(mc, us)
}
\arguments{
\item{mc}{a metacell object}

\item{us}{umi matrix}
}
\description{
The footprint is defined as the size-normalized geometric mena of the number of umis per metacells, dvided by the median over all metacells.
}
