% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.r
\name{mc_compute_n_bc}
\alias{mc_compute_n_bc}
\title{Compute distribution of cells over batches and metacell}
\usage{
mc_compute_n_bc(mc, scmat)
}
\arguments{
\item{mc}{a metacell object}

\item{scmat}{scamt object (for the metadata)}
}
\description{
Compute distribution of cells over batches and metacell
}
