% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_outliers.r
\name{mc_compute_outlier_fc}
\alias{mc_compute_outlier_fc}
\title{Compute log fold change expression of each gene given its regularized metacell expression}
\usage{
mc_compute_outlier_fc(mc, mat)
}
\arguments{
\item{mc}{metacell object (not id)}

\item{mat}{matrix object (not id)}
}
\description{
Each metacell is defiend by a total set of umis, defining an expected number of umi pere cell per gene. After regularization (1 umi per cell), the fold enrichment of the gene in each cell is computed naively. Note that the MC inferred value is currently based on linear mean, whic his highly inadequate when an outlier is present.
}
