% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.r
\name{mc_reorder}
\alias{mc_reorder}
\title{Reorder metacell data given defined order}
\usage{
mc_reorder(mc, ord)
}
\arguments{
\item{mc}{metacell object}

\item{ord}{new order metacells}
}
\description{
Reorder metacell data given defined order
}
