% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.r
\name{mc_restrict}
\alias{mc_restrict}
\title{Create a metacell object on a subset of the MCs, with all other cells becoming outliers. There is no re-normalization of mc_fp.}
\usage{
mc_restrict(mc, foc_mcs)
}
\arguments{
\item{mc}{metacell object}

\item{foc_mcs}{list of metacell to keep in the object}
}
\description{
Create a metacell object on a subset of the MCs, with all other cells becoming outliers. There is no re-normalization of mc_fp.
}
