% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.r
\name{mc_update_stats}
\alias{mc_update_stats}
\title{Compute stats over metacell and update the object}
\usage{
mc_update_stats(mc, scmat)
}
\arguments{
\item{mc}{a metacell object}

\item{scmat}{a matrix object to draw statisics from}
}
\description{
This compute statistics per metacell given an sc matirx.
}
