% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.r
\name{mcatlas_annotate_mc_by_mc2mc_projection}
\alias{mcatlas_annotate_mc_by_mc2mc_projection}
\title{Annotate query metacell with atlas by comparing query metacell gene profiles to the atlas MCs gene profiles
This will take each MC in the  query MC and find its to correlated
metacell in the reference, gemeratomg some figures along the way/}
\usage{
mcatlas_annotate_mc_by_mc2mc_projection(
  atlas_id,
  qmc_id,
  qmat_naming_type,
  new_qmc_id,
  q_gset_id = NULL,
  T_cor_gap = 1,
  nd_color = "lightgray",
  fig_cmp_dir = NULL
)
}
\arguments{
\item{atlas_id}{id of atlas object in scdb}

\item{qmc_id}{id of metacell object ina scdb}

\item{qmat_naming_type}{naming scheme of query matrix/mc}

\item{new_qmc_id}{id of recolored metacell object to save in DB (NULL will supress updating db)}

\item{q_gset_id}{query gene set id object (optional) - to restrict features used in comaprison to the intersection of atlas and query gsets}

\item{T_cor_gap}{how much gap between internal atlas mc-mc correlation and query-atlas correlation one is allowing to keep the annotation.}

\item{nd_color}{- color for metacell without an annotation}

\item{fig_cmp_dir}{name of directory to put figures per MC}
}
\description{
Annotate query metacell with atlas by comparing query metacell gene profiles to the atlas MCs gene profiles
This will take each MC in the  query MC and find its to correlated
metacell in the reference, gemeratomg some figures along the way/
}
