% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.r
\name{mcatlas_annotate_mc_by_sc2mc_projection}
\alias{mcatlas_annotate_mc_by_sc2mc_projection}
\title{Annotate query metacell with atlas by comparing query cells to atlas MCs}
\usage{
mcatlas_annotate_mc_by_sc2mc_projection(
  atlas_id,
  qmat_id,
  qmc_id,
  qmat_naming_type,
  fig_cmp_dir = NULL,
  q_gset_id = NULL,
  new_qmc_id = NULL,
  atlas_mc2d_id = NULL,
  pull_all_atlas_genes = F,
  max_entropy = 2,
  burn_cor = 0.6
)
}
\arguments{
\item{atlas_id}{id of atlas object in scdb}

\item{qmat_id}{id of metacell object ina scdb}

\item{qmc_id}{id of metacell object ina scdb}

\item{qmat_naming_type}{naming scheme of query matrix/mc}

\item{fig_cmp_dir}{name of directory to put figures per MC}

\item{q_gset_id}{query gene set id object (optional) - to restrict features used in comaprison to the intersection of atlas and query gsets}

\item{pull_all_atlas_genes}{return mean expression of all atlas genes (not only features) over query metacells based on best metacell hit of each query cell in the atlas}

\item{recolor_mc_id}{if this is specified,  the atlas colors will be projected on the query MCCs and updated to the scdb object  named recolor_mc}
}
\description{
This will take each cell in the query MC and find its to correlated
metacell in the reference, then generating figures showing detailed comparison of how each metacell in the query is distributed in the atlas, and how the pool of all cells in the query MC compare to the pool of their best match projections
}
