% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.r
\name{mcatlas_comp_adj_type_n}
\alias{mcatlas_comp_adj_type_n}
\title{Compute the summary matrix of cell type adjacencies}
\usage{
mcatlas_comp_adj_type_n(mc, graph, color_type_id)
}
\arguments{
\item{mc}{metacell object for defining colors/type per cell}

\item{cgraph}{cell-cell knn graph}

\item{color_ype_id}{named vector mapping colors to IDs}
}
\value{
summary matrix with cells in rows, types in columns
}
\description{
Compute the summary matrix of cell type adjacencies
}
