% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.r
\name{mcatlas_project_query}
\alias{mcatlas_project_query}
\title{Compute association of query umi matrix and atlas}
\usage{
mcatlas_project_query(atlas, qmat_id, qmat_naming_type, all_vs_all)
}
\arguments{
\item{atlas}{and atlas object}

\item{qmat_naming_type}{gene naming type of query}

\item{qmat}{query matrix}
}
\value{
a query processed list
}
\description{
Compute association of query umi matrix and atlas
}
