% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph_extmat.r
\name{mcell_add_cgraph_from_distmat}
\alias{mcell_add_cgraph_from_distmat}
\title{Build a cell graph using blanacing of an extrenal distance matrix}
\usage{
mcell_add_cgraph_from_distmat(d_mat, graph_id, K, balance = F,
  k_expand = 10)
}
\arguments{
\item{d_mat}{distance matrix, col and row names should be a super set of the non ignored cells for the matrix in mat_id}

\item{graph_id}{new graph id to create}

\item{K}{the guideline Knn parameter. The balanced will be constructed aiming at K edges per cell}

\item{balance}{should balancing be performed (Default False)}

\item{k_expand}{determine how much shoudl the K be expanded intially in order to find enough balanced neighbors.}

\item{mat_id}{matrix object id}
}
\description{
Build a cell graph using blanacing of an extrenal distance matrix
}
