% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph_knn.r
\name{mcell_add_cgraph_from_mat_raw_knn}
\alias{mcell_add_cgraph_from_mat_raw_knn}
\title{Build a cell graph using raw knn graph on given gene features}
\usage{
mcell_add_cgraph_from_mat_raw_knn(mat_id, gset_id, graph_id, K,
  dsamp = F)
}
\arguments{
\item{mat_id}{matrix object id}

\item{gset_id}{gset object id defining the features used for distances}

\item{graph_id}{new graph id to create}

\item{K}{the guideline Knn parameter. The balanced will be constructed aiming at K edges per cell}
}
\description{
Build a cell graph using raw knn graph on given gene features
}
