% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_names_xref_util.r
\name{mcell_add_gene_names_xref_mars10x}
\alias{mcell_add_gene_names_xref_mars10x}
\title{Create a gene name xref for heuristc map of mars/10x}
\usage{
mcell_add_gene_names_xref_mars10x(mars_mat_id, tenx_mat_id, xref_id = "DB")
}
\description{
Create a gene name xref for heuristc map of mars/10x
}
