% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat_mars.r
\name{mcell_add_mars_facs_data}
\alias{mcell_add_mars_facs_data}
\title{Creates a new matrix object from a given one by adding FACS index sorting data to matrix metadata table}
\usage{
mcell_add_mars_facs_data(
  new_mat_id,
  mat_id,
  base_dir,
  amp_batch_ids = NULL,
  force = T
)
}
\arguments{
\item{new_mat_id}{\itemize{
\item id of the new matrix in scdb
}}

\item{mat_id}{\itemize{
\item id of the original matrix to add data to
}}

\item{base_dir}{\itemize{
\item directory where raw tab-delimited index-sort data files are located. Expecting file name to be <amp_batch_id>.txt with well names as row names and antibodies as column names
}}

\item{amp_batch_ids}{\itemize{
\item if not null, upload data from amp_batch_id specified here.
}}

\item{force}{overwrite output matrix if it exist}

\item{dataset_table_fn}{\itemize{
\item path of the key table.
}}
}
\description{
#' @param mat_nm - the name of the new matrix in scdb
}
