% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_graphcov.r
\name{mcell_add_mc_from_graph}
\alias{mcell_add_mc_from_graph}
\title{Compute metacell using a native implementation of a graph cover k-means-like approach}
\usage{
mcell_add_mc_from_graph(mc_id, graph_id, mat_id, min_mc_size)
}
\arguments{
\item{mc_id}{id of metacell object to be added}

\item{graph_id}{a knn graph object id into scdb}

\item{mat_id}{a matrix object id}

\item{min_mc_size}{target minimum metacell size. This is only an approximation and smaller MC may be returned by the algorithm}
}
\description{
Compute metacell using a native implementation of a graph cover k-means-like approach
}
