% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat_batch_stat.r
\name{mcell_batch_stats}
\alias{mcell_batch_stats}
\title{generate a batch stat table for a given single cell umi matrix}
\usage{
mcell_batch_stats(scmat)
}
\arguments{
\item{scmat}{\itemize{
\item the matrix to work on
}}
}
\value{
batch stats table
}
\description{
generate a batch stat table for a given single cell umi matrix
}
