% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat_batch_stat.r
\name{mcell_calc_one_batch_stats}
\alias{mcell_calc_one_batch_stats}
\title{calc batch stats - essentially umi distribution}
\usage{
mcell_calc_one_batch_stats(scmat, batch_id)
}
\arguments{
\item{scmat}{\itemize{
\item a matrix object
}}

\item{batch_id}{\itemize{
\item the batch name to summarize
}}
}
\value{
one row data frame for merging
}
\description{
calc batch stats - essentially umi distribution
}
