% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclust_graph_cov.r
\name{mcell_coclust_from_graph_resamp}
\alias{mcell_coclust_from_graph_resamp}
\title{Compute metacell using resampling iterations of graph cover k-means-like approach}
\usage{
mcell_coclust_from_graph_resamp(
  coc_id,
  graph_id,
  min_mc_size,
  p_resamp,
  n_resamp
)
}
\arguments{
\item{graph_id}{a knn graph object id into scdb}

\item{min_mc_size}{target minimum metacell size. This is only an approximation and smaller MC may be returned by the algorithm}

\item{mc_id}{id of metacell object to be added}

\item{mat_id}{a matrix object id}

\item{resamp_n}{number of resampling iterations}
}
\description{
Compute metacell using resampling iterations of graph cover k-means-like approach
}
