% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_plot_marks.r
\name{mcell_compare_bulk}
\alias{mcell_compare_bulk}
\title{Utility plot function to compare bulk expression of two batches/metadata factors}
\usage{
mcell_compare_bulk(
  mat_id,
  meta_field = NULL,
  v1 = NULL,
  v2 = NULL,
  mc_id = NULL,
  mc1 = -1,
  mc2 = -1,
  mc_to_focus = NULL,
  genes = NULL,
  cex = 0.5,
  text_cex = 1,
  n_reg = 5,
  fig_nm = NULL,
  top_genes = 10,
  fig_h = 600,
  fig_w = 800
)
}
\arguments{
\item{mat_id}{matrix id to use}

\item{meta_field}{metadata field to select cells, if not specified than a metacell must be specified}

\item{v1}{first set metadata value}

\item{v2}{secod set metadata value}

\item{mc_id}{metacell model (must be specified if meta_field is not used)}

\item{mc1}{metacell id to define cells in first set}

\item{mc2}{metacell id to define cells in second set}

\item{mc_to_focus}{metacell IDs to restrict analysis of metadata comparison}

\item{genes}{genes to highlight (optional)}
}
\description{
Utility plot function to compare bulk expression of two batches/metadata factors
}
