% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph_knn.r
\name{mcell_gen_cell_cor_gset}
\alias{mcell_gen_cell_cor_gset}
\title{Compute a cell cell correlation matrix using features defined by a gene set}
\usage{
mcell_gen_cell_cor_gset(mat_id, gset_id, dsamp = F)
}
\arguments{
\item{mat_id}{matrix object id}

\item{gset_id}{gset object id defining the features used for distances}
}
\description{
Compute a cell cell correlation matrix using features defined by a gene set
}
