% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset_mc.r
\name{mcell_gset_add_gene}
\alias{mcell_gset_add_gene}
\title{adding genes to a gene set}
\usage{
mcell_gset_add_gene(gset_id, genes, subset_id = 1)
}
\arguments{
\item{gset_id}{the basic gene set}

\item{genes}{names of genes to add}

\item{subset_id}{id of subset for adding the genes, 1 by defulat}
}
\description{
adding genes to a gene set
}
