% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset_clust.r
\name{mcell_gset_filt_clusts}
\alias{mcell_gset_filt_clusts}
\title{add new gene set based on an existing one with filtering specific clusters}
\usage{
mcell_gset_filt_clusts(gset_id, filt_clusts, new_id)
}
\arguments{
\item{gset_id}{current gene set object. Only the list of genes will be used}

\item{filt_clusts}{list of clusters for filtering}

\item{new_id}{id of newly generated gene set}
}
\description{
add new gene set based on an existing one with filtering specific clusters
}
