% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset_feat_select.r
\name{mcell_gset_filter_cov}
\alias{mcell_gset_filter_cov}
\title{gnereate/filter gene features from coverage threshold in gstat table}
\usage{
mcell_gset_filter_cov(gstat_id, gset_id, T_tot, T_top3, force_new = F)
}
\arguments{
\item{gstat_id}{the ID of the gstat object to use}

\item{gset_id}{if this exists, filter_varmean will restrict hte current genes in the set with genes matching the var mean threshold, if not, it will generate a new gene sets object with one set including all high variance genes}

\item{T_tot}{total down sampled coverage threhsold}

\item{T_top3}{threshold value for the third highest umi count for the gene (only genes with top3>T_top3 are used)}
}
\description{
gnereate/filter gene features from coverage threshold in gstat table
}
