% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset_feat_select.r
\name{mcell_gset_filter_szcor}
\alias{mcell_gset_filter_szcor}
\title{gnereate/filter gene features from statistics on correlation with umi count}
\usage{
mcell_gset_filter_szcor(gstat_id, gset_id, T_szcor, force_new = F)
}
\arguments{
\item{gstat_id}{the ID of the gstat object to use}

\item{gset_id}{if this exists, filter_varmean will restrict hte current genes in the set with genes matching the var mean threshold, if not, it will generate a new gene sets object with one set including all high variance genes}

\item{T_szcor}{upper limit on normalized sz_cor (low values mark interesting gene features). If you use this, consider values around -0.1 - but evaluate carefully your decision using the gstat empirical data}
}
\description{
gnereate/filter gene features from statistics on correlation with umi count
}
