% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset_feat_select.r
\name{mcell_gset_filter_varmean}
\alias{mcell_gset_filter_varmean}
\title{gnereate/filter gene features from gstat normalized var/mean}
\usage{
mcell_gset_filter_varmean(gstat_id, gset_id, T_vm, force_new = F)
}
\arguments{
\item{gstat_id}{the ID of the gstat object to use}

\item{gset_id}{if this exists, filter_varmean will restrict hte current genes in the set with genes matching the var mean threshold, if not, it will generate a new gene sets object with one set including all high variance genes}

\item{T_vm}{the threshold on normalized var/mean, recommended values are usually around 0.2, but this may vary with the data}
}
\description{
gnereate/filter gene features from gstat normalized var/mean
}
