% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset_mc.r
\name{mcell_gset_from_mc_markers}
\alias{mcell_gset_from_mc_markers}
\title{Generate a gene set of markers from a metacell cover}
\usage{
mcell_gset_from_mc_markers(gset_id, mc_id, filt_gset_id = NULL,
  blacklist_gset_id = NULL)
}
\arguments{
\item{gset_id}{id of gene set to generate}

\item{mc_id}{id of metacell object}

\item{filt_gset_id}{geneset to select from (null by default)}

\item{blacklist_gset_id}{geneset to exclude (null by default)}
}
\description{
This should be mostly used for visualization and highlighting key genes
}
