% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset_clust.r
\name{mcell_gset_split_by_dsmat}
\alias{mcell_gset_split_by_dsmat}
\title{split gene set into several modules using clustering of genes by correlation over cells downsampled umi vector}
\usage{
mcell_gset_split_by_dsmat(gset_id, mat_id, K, force = F)
}
\arguments{
\item{gset_id}{current gene set object. Only the list of genes will be used}

\item{mat_id}{scRNA matrix to use. We assume genes in the set are represented in the matrix, but if force is T we allow filtering of genes by represented matrix entry in addtion to clsutering}

\item{K}{how many gene modules to generate (kmeans parameter)}

\item{force}{if this is T, we allow clustering usign matrix in which not all genes are represnted. This will override the existing gset will a subset of the genes!}

\item{(ENV)}{scm_k_nonz_exp - to define transformation of UMIs}
}
\description{
split gene set into several modules using clustering of genes by correlation over cells downsampled umi vector
}
