% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{mcell_mat_ignore_genes}
\alias{mcell_mat_ignore_genes}
\title{Generate a new matrix object with a given ignore gene list}
\usage{
mcell_mat_ignore_genes(new_mat_id, mat_id, ig_genes, reverse = F)
}
\arguments{
\item{new_mat_id}{id of matrix in scdb}

\item{mat_id}{existing matrix}

\item{reverse}{set this to true to focus on genes instead of ignore them. False by default}

\item{ig_gene}{gene names to ignore}
}
\description{
Generate a new matrix object with a given ignore gene list
}
