% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat_gcors.r
\name{mcell_mat_rpt_cor_anchors}
\alias{mcell_mat_rpt_cor_anchors}
\title{Find and report naive gene gene correlation over umi matrix}
\usage{
mcell_mat_rpt_cor_anchors(mat_id, gene_anchors, gene_anti = c(),
  cor_thresh, tab_fn, sz_cor_thresh = NA, downsample_n = NA,
  cells = NULL)
}
\arguments{
\item{mat_id}{matrix to use}

\item{gene_anchors}{genes that will be used as anchors for computing correlation against all other genes}

\item{gene_anti}{genes that will be used as negative anchors - high correlation to these will be considered negatively}

\item{cor_thresh}{correlation threshold (on downsampled data) - the table will include genes with maximal correlation higher than the threshold, AND, maximal correlation for negative anchor lower than for positive nachors}

\item{tab_fn}{file name of tabular report}
}
\description{
The generated table should be the basis for manual curation usually, and can be used to generate supervised gene sets, or a basis for them
}
