% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc2d_force_knn.r
\name{mcell_mc2d_force_knn}
\alias{mcell_mc2d_force_knn}
\title{Metacell layout using force directed projection of a low degree mc graph}
\usage{
mcell_mc2d_force_knn(mc2d_id, mc_id, graph_id, ignore_mismatch = F,
  symmetrize = F, ignore_edges = NULL, use_umap_on_mc_egc = F,
  gset_for_umap = NULL, uconf = NULL, umap_exclude_genes = NULL)
}
\arguments{
\item{mc2d_id}{2d object to add}

\item{mc_id}{meta cell id to work with}

\item{graph_id}{graph_id of the similarity graph on cells from the metacell}

\item{ignore_mismatch}{}

\item{symetrix}{should the mc confusion matrix be symmetrized before computing layout?}
}
\description{
Metacell layout using force directed projection of a low degree mc graph
}
