% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc2d_force_knn.r
\name{mcell_mc2d_force_knn_on_cells}
\alias{mcell_mc2d_force_knn_on_cells}
\title{Compute cells 2d coordinates based on the mc graph when the mc coordinates are supplied externally}
\usage{
mcell_mc2d_force_knn_on_cells(mc2d_id, mc_id, graph_id, mc_xy,
  ignore_mismatch = F)
}
\arguments{
\item{mc2d_id}{2d object to add}

\item{mc_id}{meta cell id to work with}

\item{graph_id}{graph_id of the similarity graph on cells from the metacell}

\item{mc_xy}{pre-defined metacells coordinates (so only coordinates of cells will be computed). Data frame with 2 columns named mc_x and mc_y (x and y coodinates, respectively).}

\item{ignore_mismatch}{}
}
\description{
Compute cells 2d coordinates based on the mc graph when the mc coordinates are supplied externally
}
