% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc2d_plot.r
\name{mcell_mc2d_plot_by_factor}
\alias{mcell_mc2d_plot_by_factor}
\title{Plot mc+cells using pre-defined mc colorization, breakdown by given metadata field (e.g. patient)}
\usage{
mcell_mc2d_plot_by_factor(mc2d_id, mat_id, meta_field,
  meta_data_vals = NULL, single_plot = T, filter_values = NULL,
  filter_name = NULL, ncols = NULL, neto_points = F)
}
\arguments{
\item{mc2d_id}{mc2d object to use for plot}

\item{mat_id}{mat object matching mc2d_id that contains the cells metadata information}

\item{meta_field}{field name (in mat cell_metadata slot) to split cells by}

\item{meta_data_vals}{actual named vector (names are cells, values are the metadata to factor on), if this is not null, the meta_field is not used}

\item{single_plot}{output all panels in a single plot or plot per panel (T)}

\item{filter_values}{to filter meta_field values by (NULL)}

\item{filter_name}{name to add to plots (NULL)}

\item{ncols}{number of panels in column (if single_plot is true), automatically calculate if NULL}

\item{neto_points}{plot without a box and a title (relevant if single_plot=F)}
}
\description{
Plot mc+cells using pre-defined mc colorization, breakdown by given metadata field (e.g. patient)
}
