% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc2d_plot.r
\name{mcell_mc2d_plot_gene}
\alias{mcell_mc2d_plot_gene}
\title{Plot the (log2) metacell footprint value of the selected gene on the 2d projection}
\usage{
mcell_mc2d_plot_gene(
  mc2d_id,
  gene,
  show_mc_ids = F,
  show_legend = T,
  neto_points = F,
  max_lfp = NA,
  min_lfp = NA,
  color_cells = F,
  mat_ds = NULL,
  zero_sc_v = 0,
  one_sc_v = 1,
  two_sc_v = 2,
  base_dir = NULL,
  filt_mc = NULL
)
}
\arguments{
\item{mc2d_id}{mc2d object to use for plot}

\item{gene}{gene name to plot}

\item{show_mc_ids}{plot metacell ids (false by default)}

\item{show_legend}{plot color bar legend (true by default)}

\item{neto_points}{do not plot box, title and legend}

\item{color_cells}{(F default) should cells be colored by UMIs}

\item{mat_ds}{downsampled matrix for coloring cells}

\item{zero_sc_v}{num of umis to consdier as 0 for cell coloring (def 0)}

\item{one_sc_v}{num of umis to consdier as 1 for cell coloring (def 1)}

\item{base_dir}{override the default base directory if not null}

\item{filt_mc}{(defulat is NULL) - factor to determine which metacells to plot}

\item{tw_sc_v}{num of umis to consdier as 1 for cell coloring (def 2)}
}
\description{
Plot the (log2) metacell footprint value of the selected gene on the 2d projection
}
