% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc2d.r
\name{mcell_mc2d_rotate}
\alias{mcell_mc2d_rotate}
\title{Rotatae/invert projection}
\usage{
mcell_mc2d_rotate(mc2d_id, alpha = 0, flipx = F, flipy = F)
}
\arguments{
\item{mc2d_id}{new id of 2d object to add}

\item{alpha}{angle in degrees}

\item{flipx}{should flip the X over the mean(x)}

\item{flipy}{should flip the Y over the mean(Y)}
}
\description{
transform all coordainte by a rotation and inversion
}
