% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.r
\name{mcell_mc_add_annot}
\alias{mcell_mc_add_annot}
\title{Update metacell annotation}
\usage{
mcell_mc_add_annot(mc_id, annots)
}
\arguments{
\item{mc}{a metacell object}

\item{annot}{annotation vectors}
}
\description{
Update metacell annotation
}
