% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.r
\name{mcell_mc_add_color}
\alias{mcell_mc_add_color}
\title{Update metacells  colors}
\usage{
mcell_mc_add_color(mc_id, colors)
}
\arguments{
\item{colors}{a vector of colors per metacell}

\item{mc}{a metacell object}
}
\description{
Update metacells  colors
}
