% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_confusion.r
\name{mcell_mc_cell_homogeneity}
\alias{mcell_mc_cell_homogeneity}
\title{Compute cell homogeneity - the fraction of intra mc edges per cell}
\usage{
mcell_mc_cell_homogeneity(mc_id, graph_id, ignore_mis = F, T_w = 0)
}
\arguments{
\item{mc_id}{meta cell object}

\item{graph_id}{graph object}
}
\description{
Compute cell homogeneity - the fraction of intra mc edges per cell
}
