% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_confusion.r
\name{mcell_mc_confusion_mat}
\alias{mcell_mc_confusion_mat}
\title{Compute confusion matrix on metacells}
\usage{
mcell_mc_confusion_mat(mc_id, graph_id, K)
}
\arguments{
\item{mc_id}{meta cell object}

\item{graph_id}{graph object}

\item{K}{Knn parameter for filtering graph edges}
}
\description{
Compute confusion matrix on metacells
}
