% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_apx_qc.r
\name{mcell_mc_edge_density}
\alias{mcell_mc_edge_density}
\title{for each MC we compute mean and variance of the intra-MC edge density}
\usage{
mcell_mc_edge_density(mc_id, graph_id, mc2d_id = NA)
}
\arguments{
\item{mc_id}{MC cover to analyze}

\item{cgraph_id}{which cgraph to study}
}
\description{
for each MC we compute mean and variance of the intra-MC edge density
}
