% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_export_tab.r
\name{mcell_mc_export_tab}
\alias{mcell_mc_export_tab}
\title{Output genes cell modules footprint matrix with metadata on the cell modules}
\usage{
mcell_mc_export_tab(mc_id, gstat_id = NULL, T_gene_tot = 50, T_fold = 1,
  metadata_fields = c("batch_set_id"))
}
\arguments{
\item{mc_id}{meta cell object id}

\item{gstat_id}{gstat object id}

\item{T_gene_tot}{minimal number of umis to include gene (to reduce table size)}

\item{T_fold}{threshold for maximal fold change over metacells}

\item{metadata_fields}{metadata field names to be used as factors and breakdown over mcs}
}
\value{
Nothing. But save a table in the figure directory that can be loaded to excel
}
\description{
Output genes cell modules footprint matrix with metadata on the cell modules
}
