% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_from_coclust.r
\name{mcell_mc_from_coclust_balanced}
\alias{mcell_mc_from_coclust_balanced}
\title{build a metacell cover from co-clust data through filtering un-balanced edges
and running graph cover}
\usage{
mcell_mc_from_coclust_balanced(
  mc_id,
  coc_id,
  mat_id,
  K,
  min_mc_size,
  alpha = 2
)
}
\arguments{
\item{mc_id}{Id of new metacell object}

\item{coc_id}{cocluster object to use}

\item{mat_id}{mat object to use when building the mc object}

\item{K}{\itemize{
\item this will
}}

\item{min_mc_size}{minimum mc size for graph cov}

\item{alpha}{the threshold for filtering edges by their coclust weight is alpha * (Kth highest coclust on either node1 or node2)}
}
\description{
build a metacell cover from co-clust data through filtering un-balanced edges
and running graph cover
}
