% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_from_coclust.r
\name{mcell_mc_from_coclust_hc}
\alias{mcell_mc_from_coclust_hc}
\title{build a metacell cover from a co-clust object using a simple hclust approach}
\usage{
mcell_mc_from_coclust_hc(mc_id, coc_id, mat_id, K, force = F)
}
\arguments{
\item{mc_id}{Id of new metacell object}

\item{coc_id}{cocluster object to use}

\item{mat_id}{mat object to use when building the mc object}

\item{K}{number of clusters to generate from the coclust hclust. Look at the coclust plot to detemrine this. But make sure you keep the clusters at reasonable size (e.g. <100) for use as metacells.}

\item{force}{set this to true to overide size limits on hclust}
}
\description{
build a metacell cover from a co-clust object using a simple hclust approach
}
