% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_hierarchy.r
\name{mcell_mc_hclust_confu}
\alias{mcell_mc_hclust_confu}
\title{comput mc hierarchucal clustering using the normalized confusion matrix}
\usage{
mcell_mc_hclust_confu(mc_id, graph_id, confu = NULL)
}
\arguments{
\item{mc_id}{metacell id}

\item{graph_id}{cgraph to define edges and impose them on the metacells}

\item{confu}{a confusion matrix derived from e.g. mcell_mc_confusion_matm or mcell_mc_coclust_confusion_mat}
}
\description{
comput mc hierarchucal clustering using the normalized confusion matrix
}
