% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_plot_marks.r
\name{mcell_mc_plot_gg}
\alias{mcell_mc_plot_gg}
\title{Utility plot function to compare two genes based on mc_fp}
\usage{
mcell_mc_plot_gg(
  mc_id,
  g1,
  g2,
  cex = 2,
  text_cex = 0.5,
  mc_colors = NULL,
  fig_fn = NULL,
  cell_md = NULL,
  md_mode = "filter",
  use_egc = F,
  main = NULL,
  cex.main = 1,
  xlim = NULL,
  ylim = NULL,
  add_grid = F,
  e_gc_eps = 1e-05,
  mc_filt = NULL,
  md = NULL,
  md_regexp = NULL,
  md_field = NULL
)
}
\arguments{
\item{mc_id}{metacell id in scdb}

\item{g1}{gene name}

\item{g2}{gene name}

\item{cex}{point size}

\item{text_cex}{text size (default 0.5)}

\item{mc_colors}{coloring vector per MC, default is NULL in which case we use the MC colors}

\item{fig_fn}{file name for the figure (optional - default is current device)}

\item{cell_md}{a factor/logical vector defining which cells are in the focus. If this is unspecified all cells/metacells will be included}

\item{md_mode}{one of "filter", "highlight"}

\item{use_egc}{plot log of absolute expression instead of the log of relative expression (normalized to the median over all MCs)}

\item{mc_filt}{potential facator filtering metacell ids (default NULL)}

\item{md}{metadata to use for highlighting/filtering}

\item{md_regexp}{regular expression to filter/highlight metadata}

\item{md_field}{metadata field to use}
}
\description{
Utility plot function to compare two genes based on mc_fp
}
