% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_hierarchy.r
\name{mcell_mc_plot_hierarchy}
\alias{mcell_mc_plot_hierarchy}
\title{plot super strucutre: super clust mc footprint, and selected genes}
\usage{
mcell_mc_plot_hierarchy(mc_id, graph_id, mc_order, sup_mcs, width, height,
  fig_fn = NULL, min_nmc = 2)
}
\arguments{
\item{mc_id}{id of metacell object}

\item{graph_id}{id of graph for (Re-) constructing confusion matrix}

\item{mc_order}{the mc ordering (e.g., hc$order using the output of mcell_mc_hclust_confu)}

\item{sup_mcs}{the list you get from mcell_mc_hierarchy (for now)}

\item{width}{width of figure in pixels}

\item{fig_fn}{figure name (NULL will create a figure named XX_supmc_confu in the figure directory)}

\item{min_nmc}{minimal number of mc in supmc set, smaller gruops will not be plotted}

\item{heigh}{heigh of figure in pixels}
}
\description{
plot super strucutre: super clust mc footprint, and selected genes
}
