% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_plot_marks.r
\name{mcell_mc_plot_subheats}
\alias{mcell_mc_plot_subheats}
\title{PLot a series of heat maps to describe metacell groups}
\usage{
mcell_mc_plot_subheats(mc_id, mat_id, lateral_gset_id)
}
\arguments{
\item{mc_id}{metacell id in scdb}

\item{mat_id}{matrix object}

\item{lateral_gset_id}{genes to be marked as lateral}
}
\description{
PLot a series of heat maps to describe metacell groups
}
