% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_plot_marks.r
\name{mcell_mc_plot_submc_marks}
\alias{mcell_mc_plot_submc_marks}
\title{plot a marker heat map for a subset of metacells - selecting relevant genes for separation}
\usage{
mcell_mc_plot_submc_marks(
  mc_id,
  mat_id,
  foc_mcs,
  fig_fn,
  n_max_marks = 10,
  add_genes = NULL,
  zero_median = F,
  h = 800,
  w = 600
)
}
\arguments{
\item{mc_id}{id of metacell object ina scdb}

\item{mat_id}{matrix object to us (default is the mc_id - assuming they use the same ID), if some cells in mc@mc are missing from the matrix, the function will generate an error}

\item{foc_mcs}{list of metacell ids to focus on}

\item{fig_fn}{file name for the figure (if null it will be call heat_marks in the fig directory)}

\item{zero_median}{should median expression be used for trimming low umi counts}

\item{h}{height}

\item{w}{height}

\item{n_max_mark}{number of markers for separation on each metacell}

\item{ddd_genes}{specific list of genes to add to the marker list}
}
\description{
plot a marker heat map for a subset of metacells - selecting relevant genes for separation
}
