% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_outliers.r
\name{mcell_mc_screen_outliers_1gene_fold}
\alias{mcell_mc_screen_outliers_1gene_fold}
\title{Simple screen for outlier cells in a metacell cover, finding genes with overly high expression given their metacell mean}
\usage{
mcell_mc_screen_outliers_1gene_fold(new_mc_id, mc_id, mat_id, T_lfc)
}
\arguments{
\item{new_mc_id}{id of new metacell to create}

\item{mc_id}{ID of metacell in scdb}

\item{mat_id}{matrix object to use (should be compatible with mc_id)}

\item{T_lfc}{maximal log2 fold change over the expected to be considered as an outlier.}
}
\description{
Simple screen for outlier cells in a metacell cover, finding genes with overly high expression given their metacell mean
}
