% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.r
\name{mcell_mc_split_by_color_group}
\alias{mcell_mc_split_by_color_group}
\title{Splits input metacell object into sub-objects by color group, naming the new metacells <mc_id>_submc_<group>}
\usage{
mcell_mc_split_by_color_group(mc_id, mat_id, min_cells_per_sub_mc = 500)
}
\arguments{
\item{mc_id}{input mc object}

\item{mat_id}{mat object corresponsing to mc_id}

\item{min_cells_per_sub_mc}{minimum number of cells per group required to create a new metacell}
}
\description{
Splits input metacell object into sub-objects by color group, naming the new metacells <mc_id>_submc_<group>
}
