% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_outliers.r
\name{mcell_mc_split_filt}
\alias{mcell_mc_split_filt}
\title{Split and filter metacells using dbscan and outlier gene detection}
\usage{
mcell_mc_split_filt(new_mc_id, mc_id, mat_id, T_lfc, plot_mats = T,
  dirichlet = F)
}
\arguments{
\item{new_mc_id}{id of metacells to create}

\item{mc_id}{id of source metacell object}

\item{mat_id}{id of umi matrix}

\item{plot_mats}{if this is true, a heatmap will be generated for each metacell}

\item{dirichlet}{if this is true, the distance metric to be used for splitting is dirichlet on the downsampled umi counts. Otherwise (and by default), the correlation between log(1+umi) of the downsampled matrix will be used. Note that downsampling is done given the miminum depth of each metacell, seperately.}
}
\description{
Running intra-metacell clustering on each metacell, identifying splits if existing. Also remove cells with outlier expression from each metacell. Note that this can in many case eliminate doublets as outliers - but may be insffucieint as doublets may constitute homogeneous metacells in large datasets.
}
