% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{mcell_merge_mats}
\alias{mcell_merge_mats}
\title{ncell_merge_mats: Merge two matrix using their ids in scdb. See scm_merge_mats for details on batch management and policies on missing genes}
\usage{
mcell_merge_mats(id1, id2, new_id)
}
\arguments{
\item{id1}{- first matrix to merge}

\item{id2}{- second matrix to merge}

\item{new_id}{- id of matrix to generate}
}
\description{
ncell_merge_mats: Merge two matrix using their ids in scdb. See scm_merge_mats for details on batch management and policies on missing genes
}
