% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph_knn.r
\name{mcell_mgraph_knn}
\alias{mcell_mgraph_knn}
\title{Compute metacell manifod graph using the confusion matrix of balanced K-nn between individual cells projected on metacells}
\usage{
mcell_mgraph_knn(
  mgraph_id,
  mc_id,
  graph_id = NULL,
  mctnetwork_id = NULL,
  symmetrize = F,
  ignore_mismatch = F,
  ignore_edges = NULL
)
}
\arguments{
\item{mgraph_id}{id of new object}

\item{mc_id}{meta cell id to work with}

\item{graph_id}{graph_id of the similarity graph on cells from the metacell, in case confusion should be computed from similarities.}

\item{mctnetwork_id}{id of a network object, in case confusion should be comuted from flows.}

\item{ignore_mismatch}{try if the cgraph id can be partially overlapping with the metacell object - false by defualt and should be kept this way}

\item{ignore_edges}{provide a data frame with mc1,mc2 pairs of edges to delete manually}

\item{symetrize}{should the mc confusion matrix be symmetrized before computing layout?}
}
\description{
Compute metacell manifod graph using the confusion matrix of balanced K-nn between individual cells projected on metacells
}
