% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph_logistic.r
\name{mcell_mgraph_logistic}
\alias{mcell_mgraph_logistic}
\title{Compute metacell manifold graph using logistic distances and balanced K-nn}
\usage{
mcell_mgraph_logistic(
  mgraph_id,
  mc_id,
  feats_gset,
  ignore_edges = NULL,
  feats_exclude = NULL,
  logist_loc = 1,
  logist_scale = 0.2,
  logist_eps = 0.00004,
  max_d_fold = 3
)
}
\arguments{
\item{mgraph_id}{id of new object}

\item{mc_id}{meta cell id to work with}

\item{feats_gset}{gene set name for use for computing distances}

\item{ignore_edges}{provide a data frame with mc1,mc2 pairs of edges to delete manually}

\item{feats_exclude}{list of genes to exclude from the features gene set}

\item{logist_loc}{the "location" parametr of the logistic function used to determine parametric distances between metacelles}

\item{logist_scale}{the "lscale" parametr of the logistic function used to determine parametric distances between metacelles}
}
\description{
Compute metacell manifold graph using logistic distances and balanced K-nn
}
