% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc2d.r
\name{mcell_new_mc2d}
\alias{mcell_new_mc2d}
\title{Generate a new metacell in scdb}
\usage{
mcell_new_mc2d(mc2d_id, mc_id, mc_x, mc_y, sc_x, sc_y, graph)
}
\arguments{
\item{mc2d_id}{new id of 2d object to add}

\item{mc_id}{id of scdb meta cell object we link to}

\item{mc_x}{x coordinates of metacells}

\item{mc_y}{y coordinates of metacells}

\item{sc_x}{x coordinates of cells}

\item{sc_y}{y coordinates of cells}

\item{graph}{graph data frame (fields mc1,mc2 specifying edges between mc's)}
}
\description{
This constructs a meta cell cover object and puts it into scdb. It gets an MC assignment (cell->MC_ID), and a matrix, and call standard api of this class to compute the footprints.
}
