% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.r
\name{mcell_new_mcatlas}
\alias{mcell_new_mcatlas}
\title{Generate a new atlas in scdb}
\usage{
mcell_new_mcatlas(
  atlas_id,
  mat_id,
  naming_type,
  mc_id,
  graph_id,
  gset_id,
  type_annot
)
}
\arguments{
\item{mat_id}{id matrix object}

\item{naming_type}{gene name type}

\item{mc_id}{id of metacell object}

\item{graph_id}{id of cgraph object}

\item{gset_id}{id of feature gene set}

\item{type_annot}{data frame with type_id, type_name, type_color fields}

\item{edges_Knn}{the knn parameter for neighbor type distribution}
}
\description{
This constructs a meta cell atlas object
}
