% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\name{mcell_new_mctnetwork}
\alias{mcell_new_mctnetwork}
\title{Generate a new network in scdb}
\usage{
mcell_new_mctnetwork(net_id, mc_id, mgraph_id, cell_time)
}
\arguments{
\item{net_id}{id of scdb network object ot be added}

\item{mc_id}{metacell object}

\item{mgraph_id}{metacell manfold graph object}

\item{cell_time}{assigning of time (discrete) to cells}
}
\description{
This constructs a meta cell time network object from an mc and mgraph objects and assignments of cells to graphs
}
