% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat_batch_stat.r
\name{mcell_plot_batch_stats}
\alias{mcell_plot_batch_stats}
\title{plot batches stats}
\usage{
mcell_plot_batch_stats(mat_id, fig_nm = NULL)
}
\arguments{
\item{fig_nm}{output file name}

\item{batches_stats}{collected by .calc_batch_stats}

\item{cells_breakdown_field}{color #cells per batch by this field}

\item{cell_col_dict}{color dictionary for cells_breakdown_field values}
}
\description{
plot batches stats
}
