% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_plot_cross_mc.r
\name{mcell_plot_cross_mc}
\alias{mcell_plot_cross_mc}
\title{Plotting a matrix of co-occurences between two metacell covers of the same dataset}
\usage{
mcell_plot_cross_mc(mc_id1, mc_id2)
}
\arguments{
\item{mc_id1}{first mc object to use}

\item{mc_id2}{second mc object to use}
}
\description{
Plotting a matrix of co-occurences between two metacell covers of the same dataset
}
