% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset_plot_mats.r
\name{mcell_plot_gset_cor_mats}
\alias{mcell_plot_gset_cor_mats}
\title{Plot gene set correlation matrices given a an scamt. See version using metacells for potentially more robust behavior. This is used to detemrine initial feature selectio (e.g. filtering biologically irrelevant gene modules)}
\usage{
mcell_plot_gset_cor_mats(gset_id, scmat_id, downsamp = T)
}
\description{
Plot gene set correlation matrices given a an scamt. See version using metacells for potentially more robust behavior. This is used to detemrine initial feature selectio (e.g. filtering biologically irrelevant gene modules)
}
