% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstats_report.r
\name{mcell_plot_gstats}
\alias{mcell_plot_gstats}
\title{plot gene/feature statistics}
\usage{
mcell_plot_gstats(gstat_id, gset_id = NULL, fig_dir = NULL)
}
\arguments{
\item{gstat_id}{id of gene state object}

\item{gset_id}{optional gene set, will be used to mark select genes if specified}

\item{fig_dir}{optional base dir (default is to use the std fig dir)}
}
\description{
plot gene/feature statistics
}
