% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_outliers.r
\name{mcell_plot_outlier_heatmap}
\alias{mcell_plot_outlier_heatmap}
\title{Plot and outlier heat map.}
\usage{
mcell_plot_outlier_heatmap(mc_id, mat_id, T_lfc, max_genes_to_plot = 500)
}
\arguments{
\item{mc}{metacell object (not id)}

\item{mat}{matrix object (not id)}
}
\description{
This select genes that are expressed in an interesting way in one of the outliers, and visualize their combinatorics as a heatmap.
}
