% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_project.r
\name{mcell_proj_on_atlas}
\alias{mcell_proj_on_atlas}
\title{Project a metacell object on a reference "atlas"}
\usage{
mcell_proj_on_atlas(
  mat_id,
  mc_id,
  atlas,
  fig_cmp_dir,
  ten2mars = T,
  gene_name_map = NULL,
  recolor_mc_id = NULL,
  plot_all_mcs = F,
  md_field = NULL,
  max_entropy = 2,
  burn_cor = 0.6
)
}
\arguments{
\item{mat_id}{id of metacell object ina scdb}

\item{mc_id}{id of metacell object ina scdb}

\item{atlas}{an object generated by mcell_gen_atlas}

\item{fig_cmp_dir}{name of directory to put figures per MC}

\item{ten2mars}{should gene mapping be attmpeted using standard MARS-seq to 10x naming convetnion}

\item{gene_name_map}{if this is dedining a named vector query_name\link{ref_name}, then the name conversion between the query and reference will be determined using it}

\item{recolor_mc_id}{if this is specified,  the atlas colors will be projected on the query MCCs and updated to the scdb object  named recolor_mc}

\item{md_field}{metadata field too use as additional factor for plotting subsets}

\item{plot_all_mc}{set this to T if you want a plot per metacell to show comparison of query pooled umi's and projected pooled umis.}
}
\description{
This will take each cell in the query object and find its to correlated
metacell in the reference, then generating figures showing detailed comparison of how each metacell in the query is distributed in the atlas, and how the pool of all cells in the query MC compare to the pool of their best match projections
}
