% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat_mars.r
\name{mcell_read_multi_scmat_mars}
\alias{mcell_read_multi_scmat_mars}
\title{read multiple MARS umi matrices and merge them, based on a table defining the datasets.}
\usage{
mcell_read_multi_scmat_mars(datasets_table_fn, base_dir, patch_cell_name = F)
}
\arguments{
\item{datasets_table_fn}{the index file of the MARS multi batch dataset. This is a tab delimited text file, with an arbitrary number of columns and a header line. The three mandatory fields are Amp.Batch.ID and Seq.Batch.ID, Batch.Set.ID. The first specify the ID of the batch defined by the row, and also the file name (without the .txt suffix) of the respective umi table in the base_dir provided. The second defines and ID of the sequencing batch (may be relevant for further noise cleanups beyond those done in the low-level pipeline). The third id group different batches into sets for downstream analysis (e.g. QC and more). This is specifically geared toward the output of the std MARS pipeline circa 2014...(bummer).}

\item{base_dir}{defines the umitab directory}
}
\description{
read multiple MARS umi matrices and merge them, based on a table defining the datasets.
}
