% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat_ccor.r
\name{mcell_scmat_plot_cmp_kcor}
\alias{mcell_scmat_plot_cmp_kcor}
\title{Analyze cell cell cor}
\usage{
mcell_scmat_plot_cmp_kcor(mat_id, gset_id)
}
\arguments{
\item{mat_id}{matrix to use}

\item{gset_id}{features to use for computing cors}

\item{n_dsamp}{number of umis to for downsampling. If this is -1, no dsamp will be performed}
}
\description{
Analyze cell cell cor
}
