% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork_clust.r
\name{mctnetwork_clust_flows}
\alias{mctnetwork_clust_flows}
\title{Clustring MCs based on the flow graph}
\usage{
mctnetwork_clust_flows(mct_id, K)
}
\arguments{
\item{mct_id}{mct network object}
}
\value{
A matrix on metacells and time with mean umi fraction over the gene module
}
\description{
Clustring MCs based on the flow graph
}
