% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\name{mctnetwork_comp_propagation}
\alias{mctnetwork_comp_propagation}
\title{Compute matrices for forward and backwrd propagation using the flows}
\usage{
mctnetwork_comp_propagation(mct)
}
\arguments{
\item{mct}{mct network object}
}
\value{
an updated mct object
}
\description{
Compute matrices for forward and backwrd propagation using the flows
}
