% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\name{mctnetwork_gen_e_gmt_p}
\alias{mctnetwork_gen_e_gmt_p}
\title{This generate two large matrices showing expression of genes per metacell ad time point, as well as the expression of the gene in inferred ancestral states given the flow model}
\usage{
mctnetwork_gen_e_gmt_p(mct_id, mat_id, genes)
}
\arguments{
\item{mct_id}{mct network object}

\item{mat_id}{umi matrix object}

\item{genes}{list of gene names}
}
\value{
A matrix on metacells and time with mean umi fraction over the gene module
}
\description{
This generate two large matrices showing expression of genes per metacell ad time point, as well as the expression of the gene in inferred ancestral states given the flow model
}
