% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\name{mctnetwork_gen_mincost_flows}
\alias{mctnetwork_gen_mincost_flows}
\title{Compute mincost flow and update the mct objects with flows and 
inferred mc capacity per time}
\usage{
mctnetwork_gen_mincost_flows(mct, flow_tolerance = 0.01)
}
\arguments{
\item{mct}{mct network object}

\item{flow_tolerance}{how much flow we should miss per time}
}
\value{
an updated mct object
}
\description{
Compute mincost flow and update the mct objects with flows and 
inferred mc capacity per time
}
