% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\name{mctnetwork_gen_network}
\alias{mctnetwork_gen_network}
\title{Compute and update network strcuture from mc, manifodld costs}
\usage{
mctnetwork_gen_network(
  mct,
  mc_leak,
  capacity_var_factor = NULL,
  off_capacity_cost1 = 1,
  off_capacity_cost2 = 1000,
  k_norm_ext_cost = 2,
  k_ext_norm_cost = 2,
  k_ext_ext_cost = 100,
  T_cost = 1000
)
}
\arguments{
\item{mct}{mct network object}

\item{mc_leak}{the fraciton of frequency loss per mc per time step. All zero if you don't know what to expect.}
}
\value{
an updated mct object
}
\description{
Compute and update network strcuture from mc, manifodld costs
}
