% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\name{mctnetwork_get_type_flows}
\alias{mctnetwork_get_type_flows}
\title{Compute matrix of flows over cell types}
\usage{
mctnetwork_get_type_flows(mct, time, max_time)
}
\arguments{
\item{mct}{mct network object}

\item{max_time}{maximum time point}

\item{min_time}{minimum time point}
}
\value{
a list of matrices show flows from type t to t'
}
\description{
Compute matrix of flows over cell types
}
