% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfigs.r
\name{plot_color_bar}
\alias{plot_color_bar}
\title{Plot a color bar with values}
\usage{
plot_color_bar(vals, cols, fig_fn = NULL, title = "",
  show_vals_ind = NULL)
}
\arguments{
\item{vals}{to plot}

\item{cols}{colors of the values (same length as vals)}

\item{fig_fn}{if null - plot inline}

\item{title}{of legend}

\item{show_vals_ind}{logical vector - which values to show}
}
\description{
Plot a color bar with values
}
