% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.r
\name{port_cor}
\alias{port_cor}
\title{port_cor
matrix correlation wrap. Parameter selects which function to use, cor or tgs_cor. Tgs_cor can be more efficien if available}
\usage{
port_cor(x, pairwise.complete.obs = F, spearman = F, tidy = F,
  threshold = 0)
}
\arguments{
\item{x}{matrix whole columans will be analyzed}

\item{pairwise.complete.obs}{- similar to cor, if this is T, NA's will be ibe ignored and only values that are pairwise available will be used to compute the correlaiton}

\item{spearman}{TRUE if rank correlation is to be used}

\item{tidy}{if TRUE, the output will be tidy}

\item{threshold}{defining a threshold to deterine which correlation to be reported if using the tidy format}
}
\value{
the corelation matrix of the columns
}
\description{
port_cor
matrix correlation wrap. Parameter selects which function to use, cor or tgs_cor. Tgs_cor can be more efficien if available
}
