% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rescale_sparse_mat_cols}
\alias{rescale_sparse_mat_cols}
\title{Efficient version for scaling all columns in a sparse matrix}
\usage{
rescale_sparse_mat_cols(A, v_norm)
}
\arguments{
\item{A}{first matrix}

\item{B}{second matrix - should be exactly same dimension as A}
}
\description{
Efficient version for scaling all columns in a sparse matrix
}
