% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_add_coclust}
\alias{scdb_add_coclust}
\title{scdb_add_coclust - add coclust to the DB and cahce}
\usage{
scdb_add_coclust(id, coclust)
}
\arguments{
\item{id}{- id of coclust}

\item{coclust}{- coclust data frame}
}
\description{
scdb_add_coclust - add coclust to the DB and cahce
}
