% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_add_gene_names_xref}
\alias{scdb_add_gene_names_xref}
\title{scdb_add_gene_names_xref - add a gene name xref tab to the DB - will save it and cache}
\usage{
scdb_add_gene_names_xref(id = "DB", gene_names_xref)
}
\arguments{
\item{id}{- xref id. Will return null if does not exist}

\item{mat}{- xref object}
}
\description{
scdb_add_gene_names_xref - add a gene name xref tab to the DB - will save it and cache
}
