% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_add_gset}
\alias{scdb_add_gset}
\title{scdb_add_gset - add gset to the DB and cahce}
\usage{
scdb_add_gset(id, gset)
}
\arguments{
\item{id}{\itemize{
\item id of gset
}}

\item{gset}{\itemize{
\item gset data frame
}}
}
\description{
scdb_add_gset - add gset to the DB and cahce
}
