% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_add_gstat}
\alias{scdb_add_gstat}
\title{scdb_add_gstat - add gstat to the DB and cahce}
\usage{
scdb_add_gstat(id, gstat)
}
\arguments{
\item{id}{\itemize{
\item id of gstat
}}

\item{gstat}{\itemize{
\item gstat data frame
}}
}
\description{
scdb_add_gstat - add gstat to the DB and cahce
}
