% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_add_mat}
\alias{scdb_add_mat}
\title{scdb_add_mat - add amatrix to the DB - will save it and cache}
\usage{
scdb_add_mat(id, mat)
}
\arguments{
\item{id}{-  matrix id. Will return null if does not exist}

\item{mat}{- matrix object}
}
\description{
scdb_add_mat - add amatrix to the DB - will save it and cache
}
