% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_add_mc}
\alias{scdb_add_mc}
\title{scdb_add_mc - add mc to the DB and cahce}
\usage{
scdb_add_mc(id, mc)
}
\arguments{
\item{id}{\itemize{
\item id of mc
}}

\item{mc}{\itemize{
\item mc data frame
}}
}
\description{
scdb_add_mc - add mc to the DB and cahce
}
