% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_add_mc2d}
\alias{scdb_add_mc2d}
\title{scdb_add_mc2d - add mc2d to the DB and cahce}
\usage{
scdb_add_mc2d(id, mc2d)
}
\arguments{
\item{id}{- id of mc2d}

\item{mc2d}{- mc2d data frame}
}
\description{
scdb_add_mc2d - add mc2d to the DB and cahce
}
