% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_add_mcatlas}
\alias{scdb_add_mcatlas}
\title{scdb_add_mcatlas - add mcatlas to the DB and cahce}
\usage{
scdb_add_mcatlas(id, mcatlas)
}
\arguments{
\item{id}{- id of mcatlas}

\item{mcatlas}{- mcatlas object to add}
}
\description{
scdb_add_mcatlas - add mcatlas to the DB and cahce
}
