% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_add_mctnetwork}
\alias{scdb_add_mctnetwork}
\title{scdb_add_mctnetwork - add mctnetwork to the DB and cahce}
\usage{
scdb_add_mctnetwork(id, mctnetwork)
}
\arguments{
\item{id}{- id of mctnetwork}

\item{mctnetwork}{- mctnetwork data frame}
}
\description{
scdb_add_mctnetwork - add mctnetwork to the DB and cahce
}
