% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_add_mgraph}
\alias{scdb_add_mgraph}
\title{scdb_add_mgraph - add mgraph to the DB and cahce}
\usage{
scdb_add_mgraph(id, mgraph)
}
\arguments{
\item{id}{- id of mgraph}

\item{mgraph}{- mgraph data frame}
}
\description{
scdb_add_mgraph - add mgraph to the DB and cahce
}
