% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_coclust}
\alias{scdb_coclust}
\title{scdb_coclust - get a coclust object}
\usage{
scdb_coclust(id)
}
\arguments{
\item{id}{\itemize{
\item id of coclust
}}
}
\description{
scdb_coclust - get a coclust object
}
