% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_del_cgraph}
\alias{scdb_del_cgraph}
\title{scdb_del_cgraph - del cgraph from the DB and cahce}
\usage{
scdb_del_cgraph(id)
}
\arguments{
\item{id}{\itemize{
\item id of cgraph
}}
}
\description{
scdb_del_cgraph - del cgraph from the DB and cahce
}
