% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_del_coclust}
\alias{scdb_del_coclust}
\title{scdb_del_coclust - del coclust from the DB and cahce}
\usage{
scdb_del_coclust(id)
}
\arguments{
\item{id}{\itemize{
\item id of coclust
}}
}
\description{
scdb_del_coclust - del coclust from the DB and cahce
}
