% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_del_gene_names_xref}
\alias{scdb_del_gene_names_xref}
\title{scdb_del_gene_names_xref - remove a gene names xref from the DB (not just the cache!)}
\usage{
scdb_del_gene_names_xref(id)
}
\arguments{
\item{id}{-  xref id to remove from the DB}
}
\description{
scdb_del_gene_names_xref - remove a gene names xref from the DB (not just the cache!)
}
