% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_del_gset}
\alias{scdb_del_gset}
\title{scdb_del_gset - del gset from the DB and cahce}
\usage{
scdb_del_gset(id)
}
\arguments{
\item{id}{\itemize{
\item id of gset
}}
}
\description{
scdb_del_gset - del gset from the DB and cahce
}
