% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_del_gstat}
\alias{scdb_del_gstat}
\title{scdb_del_gstat - del gstat from the DB and cahce}
\usage{
scdb_del_gstat(id)
}
\arguments{
\item{id}{\itemize{
\item id of gstat
}}
}
\description{
scdb_del_gstat - del gstat from the DB and cahce
}
