% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_del_mat}
\alias{scdb_del_mat}
\title{scdb_del_mat - remove a matrix from the DB (not just the cache!)}
\usage{
scdb_del_mat(id)
}
\arguments{
\item{id}{\itemize{
\item matrix id to remove from the DB
}}
}
\description{
scdb_del_mat - remove a matrix from the DB (not just the cache!)
}
