% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_del_mc}
\alias{scdb_del_mc}
\title{scdb_del_mc - del mc from the DB and cahce}
\usage{
scdb_del_mc(id)
}
\arguments{
\item{id}{\itemize{
\item id of mc
}}
}
\description{
scdb_del_mc - del mc from the DB and cahce
}
