% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_del_mc2d}
\alias{scdb_del_mc2d}
\title{scdb_del_mc2d - del mc2d from the DB and cahce}
\usage{
scdb_del_mc2d(id)
}
\arguments{
\item{id}{\itemize{
\item id of mc2d
}}
}
\description{
scdb_del_mc2d - del mc2d from the DB and cahce
}
