% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_del_mcatlas}
\alias{scdb_del_mcatlas}
\title{scdb_del_mcatlas - del mcatlas from the DB and cahce}
\usage{
scdb_del_mcatlas(id)
}
\arguments{
\item{id}{- id of mcatlas}
}
\description{
scdb_del_mcatlas - del mcatlas from the DB and cahce
}
