% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_del_mctnetwork}
\alias{scdb_del_mctnetwork}
\title{scdb_del_mctnetwork - del mctnetwork from the DB and cahce}
\usage{
scdb_del_mctnetwork(id)
}
\arguments{
\item{id}{- id of mctnetwork}
}
\description{
scdb_del_mctnetwork - del mctnetwork from the DB and cahce
}
