% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_del_mgraph}
\alias{scdb_del_mgraph}
\title{scdb_del_mgraph - del mgraph from the DB and cahce}
\usage{
scdb_del_mgraph(id)
}
\arguments{
\item{id}{- id of mgraph}
}
\description{
scdb_del_mgraph - del mgraph from the DB and cahce
}
