% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_gene_names_xref}
\alias{scdb_gene_names_xref}
\title{scdb_gene_names_xref - get gene names convertor table from db}
\usage{
scdb_gene_names_xref(id = "DB")
}
\arguments{
\item{id}{-  id of convertor, default is "DB"}
}
\description{
scdb_gene_names_xref - get gene names convertor table from db
}
