% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_gstat}
\alias{scdb_gstat}
\title{scdb_gstat - get a gstat data frame. If it is missing and the id match an existing matrix, a gstat will be gerated for this matrix and added to scdb}
\usage{
scdb_gstat(id)
}
\arguments{
\item{id}{- id of gstat}
}
\description{
scdb_gstat - get a gstat data frame. If it is missing and the id match an existing matrix, a gstat will be gerated for this matrix and added to scdb
}
