% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_init}
\alias{scdb_init}
\title{Initializing scdb}
\usage{
scdb_init(base_dir, force_reinit = F)
}
\description{
This initialize scdb to a certain directory. The object will then allow loading. chaching and saving of different objects, including matrix, gene statictics, cell graphs, metacell covers, and more
}
\examples{
\dontrun{
# we first initialize a db object
scdb_init("workdir")
scdb_ls("mat")
}

}
