% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_ls}
\alias{scdb_ls}
\title{List all object of a given type from the current scdb}
\usage{
scdb_ls(objt, regexp = ".")
}
\arguments{
\item{objt}{- the object type}
}
\description{
List all object of a given type from the current scdb
}
