% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_ls_loaded}
\alias{scdb_ls_loaded}
\title{scdb_ls_loaded - list loaded object of a certain type}
\usage{
scdb_ls_loaded(objt)
}
\arguments{
\item{objt}{\itemize{
\item either mat,
}}
}
\description{
scdb_ls_loaded - list loaded object of a certain type
}
