% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_mat}
\alias{scdb_mat}
\title{scdb_mat - get matrix from db, load it if needed}
\usage{
scdb_mat(id)
}
\arguments{
\item{id}{-  matrix id. Will return null if does not exist}
}
\description{
scdb_mat - get matrix from db, load it if needed
}
