% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_mctnetwork}
\alias{scdb_mctnetwork}
\title{scdb_mctnetwork - get a mctnetwork object}
\usage{
scdb_mctnetwork(id)
}
\arguments{
\item{id}{- id of mctnetwork}
}
\description{
scdb_mctnetwork - get a mctnetwork object
}
