% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfigs.r
\name{scfigs_dir}
\alias{scfigs_dir}
\title{Generate a standard figure dir name igven and object and figure type}
\usage{
scfigs_dir(id, type)
}
\arguments{
\item{id}{\itemize{
\item is of the object the figure relates to (E.g. mat_id)
}}

\item{type}{\itemize{
\item a string defining the figure type
}}
}
\description{
Generate a standard figure dir name igven and object and figure type
}
