% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfigs.r
\name{scfigs_init}
\alias{scfigs_init}
\title{scgfig_set_base - set base directory for metacell figures, creates it if it doesn't exist.}
\usage{
scfigs_init(base)
}
\description{
scgfig_set_base - set base directory for metacell figures, creates it if it doesn't exist.
}
