% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{scm_downsamp}
\alias{scm_downsamp}
\title{downsampl}
\usage{
scm_downsamp(umis, n)
}
\arguments{
\item{umis}{the umi matrix}

\item{n}{the constant number of molecules per cell in the output matrix}
}
\description{
generate a umi matrix in which each cell (Column) is downsampled to a fixed number of molecules. This is a ehavy an inefficient proces at the moment, so we split it into chunks and run parallel using doMC
}
