% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{scm_export_mat_to_sce}
\alias{scm_export_mat_to_sce}
\title{Export a mat object (umi matrix and metadata table) to a SingleCellExperiment object}
\usage{
scm_export_mat_to_sce(mat_id, add_log_counts = F, scale_to = 10000)
}
\arguments{
\item{mat_id}{id of the mat object to export}
}
\value{
SingleCellExpriment instance
}
\description{
Export a mat object (umi matrix and metadata table) to a SingleCellExperiment object
}
