% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{scm_ignore_cells}
\alias{scm_ignore_cells}
\title{Set ignored (i.e. blacklisted) cells}
\usage{
scm_ignore_cells(scmat, ig_cells, reverse = FALSE)
}
\arguments{
\item{scmat}{the matrix object}

\item{reverse}{false by default, if this is true the set of cells to ingore is the complement of the given list}

\item{ig_cell}{a list of cell names to ignore}
}
\description{
Given a list of cells to ignore, this will cancel any previous policy for blacklisting and remove the given cells to the ignore_mat. Downstream algorithm usually ignore these cells altogether, for any purpose including normalization. However, ignored cells can be accessed and analyzed seperately for validation/tests or when they represent some relevant biology (e.g. cell cycle)
}
