% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{scm_ignore_genes}
\alias{scm_ignore_genes}
\title{Set ignored (i.e. blacklisted) genes}
\usage{
scm_ignore_genes(scmat, ig_genes, reverse = FALSE)
}
\arguments{
\item{scmat}{the matrix object}

\item{reverse}{false by default, if this is true the set of genes to ingore is the complement of the given list}

\item{ig_gene}{a list of gene names to ignore}
}
\description{
Given a list of genes to ignore, this will cancel any previous policy for blacklisting and remove the given genes to the ignore_mat. Downstream algorithm usually ignore these genes altogether, for any purpose including normalization. However, ignored genes can be accessed and analyzed seperately for validation/tests or when they represent some relevant biology (e.g. cell cycle)
}
