% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{scm_import_sce_to_mat}
\alias{scm_import_sce_to_mat}
\title{Import a umi count matrix with metadata per cell from a SingleCellExperiment objectto a scmat object  to a SingleCellExperiment object}
\usage{
scm_import_sce_to_mat(sce, counts_slot = "counts")
}
\arguments{
\item{sce}{Input SingleCellExperiment to export}

\item{counts_slot}{name of the umi matrix slot}
}
\value{
tgScMat
}
\description{
Import a umi count matrix with metadata per cell from a SingleCellExperiment objectto a scmat object  to a SingleCellExperiment object
}
