% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{scm_new_matrix}
\alias{scm_new_matrix}
\title{Constract a tgScMat}
\usage{
scm_new_matrix(mat, cell_metadata, stat_type = "umi")
}
\arguments{
\item{mat}{A matrix containing the expression data (will be stored as a sparse matrix)}

\item{cell_metadata}{dataframe with metadata on the cells. Rownames correspond to the cell names.}

\item{stat_type}{Type of statistic. ("umi", "rpkm", "tpm" etc.). Default: "umi"}
}
\value{
A tgScMat gene expression matrix object
}
\description{
Constract a tgScMat gene expression matrix object
}
