% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{scm_sub_mat}
\alias{scm_sub_mat}
\title{Extract sub-matrix. This return a matrix object on a subset of the genes and cells.}
\usage{
scm_sub_mat(scmat, genes = NULL, cells = NULL)
}
\arguments{
\item{scmat}{A tgScMat object.}

\item{genes}{Genes range. NULL implies the entire gene set}

\item{cells}{Cells range. NULL implies the entire cell set}
}
\description{
Extract sub-matrix. This return a matrix object on a subset of the genes and cells.
}
