% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat_10x.r
\name{scmat_read_scmat_10x_custom}
\alias{scmat_read_scmat_10x_custom}
\title{Read a custom count matrix from the output of a 10x run. Batches can be stripped from the cell identifier if in BARCODE-LANE format.}
\usage{
scmat_read_scmat_10x_custom(matrix_fn, dataset_id = "NS")
}
\arguments{
\item{matrix_fn}{matrix input file name. Expecting a count matrix saved as .RDS file. Colnames represent barcodes, rownames represent custom gene annotation.}
}
\description{
Read a custom count matrix from the output of a 10x run. Batches can be stripped from the cell identifier if in BARCODE-LANE format.
}
