% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.r
\docType{class}
\name{tgCellGraph-class}
\alias{tgCellGraph-class}
\alias{tgCellGraph}
\title{Cell graph}
\description{
Representing a cell-to-cell similarity graph.
}
\section{Slots}{

\describe{
\item{\code{cell_names}}{names of cells (Ids in edges relates to these)}

\item{\code{edges}}{data frame specifiying the graph weighted edges as returned from tgs_cor_graph}

\item{\code{nodes}}{vector of node names}
}}

