% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclust.r
\docType{class}
\name{tgCoClust-class}
\alias{tgCoClust-class}
\alias{tgCoClust}
\title{Metacell colustering object}
\description{
Represent coclustering data derived by resmapling iteration and graph cover (or any other coverage/clustering heuristic)
}
\section{Slots}{

\describe{
\item{\code{graph_id}}{id of the scdb similairty graph object}

\item{\code{coclust}}{tidy format matrix with co-occurence stats}

\item{\code{n_samp}}{number of times each node was sampled}
}}

