% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset.r
\docType{class}
\name{tgGeneSets-class}
\alias{tgGeneSets-class}
\alias{tgGeneSets}
\title{Gene sets interface}
\description{
Gene sets are simply container for multiple sets of genes IDs. This is in use when extracting features for meta-cell analysis, when deriving modules of correlated genes, or when importing annotated genes or transferring them between analyses (e.g. cell cycle).
}
\section{Slots}{

\describe{
\item{\code{description}}{textual description of the gene set collection}

\item{\code{set_names}}{possible annotations of gene sets}

\item{\code{gene_set}}{set per gene}
}}

