% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlas.r
\name{tgMAtlas}
\alias{tgMAtlas}
\title{R6 class representing a cell atlas}
\description{

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mat}}{UMI matrix (class tgScMat)}

\item{\code{mc}}{Metacells (class tgMCCov)}

\item{\code{gset}}{Gene set (class tgGeneSets)}

\item{\code{mc2d}}{Cells/metacells 2D projection (class tgMC2D)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tgMAtlas$new()}}
\item \href{#method-save}{\code{tgMAtlas$save()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new tgMAtlas object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tgMAtlas$new(mat, mc = NULL, gset = NULL, mc2d = NULL, rankings = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mat, mc, gset, mc2d}}{Objects or filenames of stored objects (.Rda or .rds) that will
be used for member initalization. XXX}

\item{\code{rankings}}{A vector of metadata columns (from \code{mat}'s metdata) that can be
used for ranking projected cells.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
It is possible to load an atlas from the filesystem by passing as
\code{mat} the path of either an atlas json file or a directory holding
an atals. In the later case it is possible to pass a glob
pattern in \code{mc} that will be used to match files in the directory.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
Annotate a group of cells (umi vectors) using NN analysis

This will take each cell in the query object and find its most  correlated
cell in the reference. The type/color of the reference cell will be return.
The time tag that can be supplied for each atlas cell will be ued to generate
The distribution of best hit time and to correlate it with the distribution of
each time point within the reference it self


Save the atlas into a directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tgMAtlas$save(dname, fname = "atlas")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dname}}{Name of directory that will hold the atals. If needed, this
directory will be created if needed.}

\item{\code{fname}}{Filename pattern for the files that will be stored (e.g.
if \code{fname} is \code{"xyz"} the UMI matrix will be saved as
\code{mat.xyz.rds})}
}
\if{html}{\out{</div>}}
}
}
}
