% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc2d.r
\docType{class}
\name{tgMC2D-class}
\alias{tgMC2D-class}
\alias{tgMC2D}
\title{Meta cell 2d projection}
\description{
Representing a graph of metacells and their 2D embeddings
}
\section{Slots}{

\describe{
\item{\code{mc_id}}{id of the metacell object}

\item{\code{x}}{x coordinate of a (subset) of the metacells}

\item{\code{y}}{y coordinate of a (subset) of the metacells}

\item{\code{graph}}{datafame with the edges linking metacells}
}}

