% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.r
\docType{class}
\name{tgMCAtlas-class}
\alias{tgMCAtlas-class}
\alias{tgMCAtlas}
\title{wrap up an atlas object from mc, mc2d and matrix ids}
\arguments{
\item{mat_id}{id of metacell object ina scdb}

\item{mc_id}{id of metacell object ina scdb}

\item{gset_id}{features defining the atlas (to be sued for determing projection)}

\item{mc2d_id}{projection object id, to define the atlas 2D layout}
}
\description{
This is not doing much more than generating a list with the relevant object names bundeled. To be enhnaced at some stage.
}
