% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.r
\docType{class}
\name{tgMCCov-class}
\alias{tgMCCov-class}
\alias{tgMCCov}
\title{Meta cell cover}
\description{
Representing a meta cell cover of a given cell graph (or more generally of a scRNA data matrix)
}
\section{Slots}{

\describe{
\item{\code{mc}}{assignment of cells to metacell id}

\item{\code{cell_names}}{names of cells (all other objects use running integres relating to these names)}

\item{\code{mc_fp}}{a matrix showing for each gene (row) the relative enrichment of umis}

\item{\code{e_gc}}{a matrix showin\preformatted{                                                                                                                                                                                                                                g for each gene (row) the mean umi count (without normalizing for cell size)
}}

\item{\code{cov_gc}}{a matrix showing for each gene (row) the fraction of cells in the meta cell that are non zero for the umi}

\item{\code{n_bc}}{a matrix detemrining for each batch (row) the meta cell breakdown}

\item{\code{annots}}{names of metacells (oridnal numbers by default)}

\item{\code{colors}}{colors of metacells}

\item{\code{color_key}}{data.frame defining color per markers genes}
}}

