% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.r
\docType{class}
\name{tgMCManifGraph-class}
\alias{tgMCManifGraph-class}
\alias{tgMCManifGraph}
\title{manifold graph structure over a metacell object}
\description{
Splitting metacells over a discrete time axis, defining manifold connections and estimated flows over them
}
\section{Slots}{

\describe{
\item{\code{mc_id}}{id of the metacell object we represent as a network}

\item{\code{times_nms}}{names of the time points (Default 1:T)}

\item{\code{mc_t}}{distribution of metacells (rows) over time points (cols)}

\item{\code{mc_manifold}}{a data frame defining triplets mc1, mc2, distance.}
}}

