% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{tgs_cor_graph}
\alias{tgs_cor_graph}
\title{wrapping tgs functions to compute balanced graph from a matrix}
\usage{
tgs_cor_graph(x, knn, k_expand, k_alpha, k_beta)
}
\arguments{
\item{x}{matrix}

\item{knn}{K parameter}

\item{k_expand}{expansion parameter}

\item{k_alpha}{used to threshold blanaced scores (k_alpha\emph{K > rank_in}rank_out0}

\item{k_beta}{used to threshold in-degree on k_beta*K before final thresholding of out-degree by K}
}
\description{
wrapping tgs functions to compute balanced graph from a matrix
}
