% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\docType{class}
\name{tgScMat-class}
\alias{tgScMat-class}
\alias{tgScMat}
\title{Single cell RNA-seq matrix}
\description{
a single cell RNA matrix interface. Does not do much beyond adding
specifically gene names, cell names and labels describing the type of data.
Methods for basic stat, gene selection, clustering are implemented separately
}
\section{Slots}{

\describe{
\item{\code{mat}}{A sparse matrix containing the expression data}

\item{\code{genes}}{list of all genes. By default the matrix row names are used}

\item{\code{cells}}{list of all cells. By default the matrix col names are used}

\item{\code{stat_type}}{Type of statistic. Currently only "umi" is supported}

\item{\code{cell_metadata}}{dataframe with metadata on the cells. Rownames correspond to the cell names. By default a table consisting of only a "Batch" field is added with a trivial batch number. If the cell_metadata does not include a Batch field, it will be added with a trivial value. Additional standard fields spike_umis - the number of (filtered_) spike umis for the cell}
}}

\examples{
# we first initialize a tgScMat object
#TO ADD
}
