% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_colorize.r
\name{mc_colorize}
\alias{mc_colorize}
\title{colorize metacell using a set of prefered markers and their colors}
\usage{
mc_colorize(new_mc_id, mc_id, marker_colors = NULL, override = T)
}
\arguments{
\item{new_mc_id}{output metacell id in scdb}

\item{mc_id}{input metacell id in scdb (default: new_mc_id)}

\item{override}{if this is true, all colors are going to be set to white unless some marker match is found}

\item{marker_color}{a data frame with fields gene, group, color, priority, thresh}
}
\description{
colorize metacell using a set of prefered markers and their colors
}
