% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat_mars.r
\name{mcell_add_mars_facs_data}
\alias{mcell_add_mars_facs_data}
\title{Creates a new matrix object from a given one by adding FACS index sorting data to matrix metadata table}
\usage{
mcell_add_mars_facs_data(new_mat_id, mat_id, base_dir, amp_batch_ids = NULL,
  force = T)
}
\arguments{
\item{new_mat_id}{- id of the new matrix in scdb}

\item{mat_id}{- id of the original matrix to add data to}

\item{base_dir}{- directory where raw mars files are located (dataset table points to subdirectories of this)}

\item{amp_batch_ids}{- if not null, upload data from amp_batch_id specified here.}

\item{force}{overwrite output matrix if it exist}

\item{dataset_table_fn}{- path of the key table.}

\item{base_dir}{- directory where raw tab-delimited index-sort data files are located. Expecting file name to be <amp_batch_id>.txt with well names as row names and antibodies as column names}
}
\description{
#' @param mat_nm - the name of the new matrix in scdb
}
