% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat_10x.r
\name{mcell_import_scmat_10x}
\alias{mcell_import_scmat_10x}
\title{Load a matrix from a 10x dataset. The scdb version of mcell_read_multi_scmat_10x}
\usage{
mcell_import_scmat_10x(mat_nm, base_dir = NULL, matrix_fn = NULL,
  genes_fn = NULL, cells_fn = NULL, force = FALSE)
}
\arguments{
\item{mat_nm}{- the name of the new matrix in scdb}

\item{force}{- if true, will import from 10x files even when the matrix is present in the DB}

\item{dataset_table_fn}{- path of the key table}
}
\description{
Load a matrix from a 10x dataset. The scdb version of mcell_read_multi_scmat_10x
}
