% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat_import.r
\name{mcell_import_scmat_tsv}
\alias{mcell_import_scmat_tsv}
\title{Load a matrix from a simple dense table with genes in rows and cells in columns. First column is the gene name}
\usage{
mcell_import_scmat_tsv(mat_nm, fn, meta_fn = NULL, dset_nm = NULL,
  force = FALSE)
}
\arguments{
\item{mat_nm}{- the name of the new matrix in scdb}

\item{fn}{path of the matrix to read}

\item{meta_fn}{meta data table to read. If this is null, all cells will be onsidred as part of a single Amp.Batch and Dataset.ID}

\item{force}{- if true, will import from 10x files even when the matrix is present in the DB}
}
\description{
Load a matrix from a simple dense table with genes in rows and cells in columns. First column is the gene name
}
