% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc2d_force_knn.r
\name{mcell_mc2d_force_knn}
\alias{mcell_mc2d_force_knn}
\title{Metacell layout using force directed projection of a low degree mc graph}
\usage{
mcell_mc2d_force_knn(mc2d_id, mc_id, graph_id, skip_missing = F)
}
\arguments{
\item{mc2d_id}{2d object to add}

\item{mc_id}{meta cell id to work with}

\item{graph_id}{graph_id of the similarity graph on cells from the metacell}

\item{skip_missing}{cells}
}
\description{
Metacell layout using force directed projection of a low degree mc graph
}
