% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc2d_plot.r
\name{mcell_mc2d_plot}
\alias{mcell_mc2d_plot}
\title{Plot mc+cell graph using pre-defined mc colorization}
\usage{
mcell_mc2d_plot(
  mc2d_id,
  legend_pos = "topleft",
  plot_edges = T,
  min_edge_l = 0,
  edge_w = 1,
  short_edge_w = 0,
  colors = NULL,
  fn_suf = ""
)
}
\arguments{
\item{mc2d_id}{mc2d object to plot}

\item{legend_pos}{position of legend}

\item{plot_edges}{plot edges between metacells (true by default)}
}
\description{
Plot mc+cell graph using pre-defined mc colorization
}
