% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc2d_plot.r
\name{mcell_mc2d_plot_gene}
\alias{mcell_mc2d_plot_gene}
\title{Plot the (log2) metacell footprint value of the selected gene on the 2d projection}
\usage{
mcell_mc2d_plot_gene(mc2d_id, gene, show_mc_ids = F, show_legend = T)
}
\arguments{
\item{mc2d_id}{mc2d object to use for plot}

\item{gene}{gene name to plot}

\item{show_mc_ids}{plot metacell ids (false by default)}

\item{show_legend}{plot color bar legend (true by default)}
}
\description{
Plot the (log2) metacell footprint value of the selected gene on the 2d projection
}
