% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_plot_marks.r
\name{mcell_mc_plot_marks}
\alias{mcell_mc_plot_marks}
\title{plot a marker heat map give a metacell object}
\usage{
mcell_mc_plot_marks(mc_id, gset_id, mat_id = mc_id, fig_fn = NULL,
  lateral_gset_id = NULL, plot_cells = T)
}
\arguments{
\item{mc_id}{id of metacell object ina scdb}

\item{gset_id}{markers to plot (e.g. using mcell_gset_from_mc_markers)}

\item{mat_id}{matrix object to us (default is the mc_id - assuming they use the same ID), if some cells in mc@mc are missing from the matrix, the function will generate an error}

\item{fig_fn}{file name for the figure (if null it will be call heat_marks in the fig directory)}

\item{lateral_gset_id}{markers to plot on top (e.g. cell cycle)}

\item{plot_cells}{by defulat this is TRUE and data is shown for single cells. If this is false, than metacells fold change values will be plotted}
}
\description{
plot a marker heat map give a metacell object
}
