% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_add_cgraph}
\alias{scdb_add_cgraph}
\title{scdb_add_cgraph - add cgraph to the DB and cahce}
\usage{
scdb_add_cgraph(id, cgraph)
}
\arguments{
\item{id}{- id of cgraph}

\item{cgraph}{- cgraph data frame}
}
\description{
scdb_add_cgraph - add cgraph to the DB and cahce
}
