% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfigs.r
\name{scfigs_fn}
\alias{scfigs_fn}
\title{Generate a standard figure name igven and object and figure type}
\usage{
scfigs_fn(id, type, dir = NULL, ext = "png")
}
\arguments{
\item{id}{\itemize{
\item is of the object the figure relates to (E.g. mat_id)
}}

\item{type}{\itemize{
\item a string defining the figure type
}}

\item{dir}{\itemize{
\item output dir. If null, using figs base dir. Creates it if it doesn't exist.
}}

\item{ext}{\itemize{
\item file extension (default: png)
}}
}
\description{
Generate a standard figure name igven and object and figure type
}
